ACUMORG-16: ACU CI/CD CstCustomization util: fix issues, redesign & functionality extending
Testing full cycle app work, fix issues. Minor refactoring.

+++ ACUCustomizationUtil/Helpers/MSBuildHelper.cs
@@ -1,5 +1,7 @@
 ﻿using System.Diagnostics;
 using ACUCustomizationUtils.Configuration;
+using ACUCustomizationUtils.Configuration.ACU;
+using ACUCustomizationUtils.Extensions;
 using Spectre.Console;
 
 namespace ACUCustomizationUtils.Helpers;
@@ -11,6 +13,7 @@ public class MsBuildHelper
     private string? _msbuildPath;
     private string? _msbuildArgs;
     private string? _packageSourceBin;
+    private string? _packageName;
 
     public MsBuildHelper(IAcuConfiguration config, StatusContext ctx)
     {
@@ -24,19 +27,40 @@ public class MsBuildHelper
         _msbuildPath = GetMsbuildPath();
         _msbuildArgs = GetMsBuildArgs(_config);
         _packageSourceBin = _config.Code.PkgSourceBinDirectory!;
+        _packageName = _config.Package.PackageName;
+        
         var process = new ProcessHelper(_msbuildPath, _msbuildArgs, _ctx);
         await process.Execute();
 
-        //Copy dll`s
-        var files = Directory.GetFiles(_packageSourceBin);
-        foreach (var file in files.Where(f => f.EndsWith(".dll")))
+
+        if (Directory.Exists(_packageSourceBin))
         {
-            var sourceFile = Path.Combine(_config.Code.MsBuildTargetDirectory!, new FileInfo(file).Name);
-            if (!File.Exists(sourceFile))
-                throw new InvalidOperationException($"Source file {sourceFile} not copied!");
-            File.Copy(sourceFile, file, true);
-            if (!File.Exists(file)) throw new InvalidOperationException($"Target file {file} not copied!");
+            //Copy dll`s
+            var files = Directory.GetFiles(_packageSourceBin);
+            foreach (var file in files.Where(f => f.EndsWith(".dll")))
+            {
+                var sourceFile = Path.Combine(_config.Code.MsBuildTargetDirectory!, new FileInfo(file).Name);
+                if (!File.Exists(sourceFile))
+                    throw new InvalidOperationException($"Source file {sourceFile} not copied!");
+                File.Copy(sourceFile, file, true);
+                if (!File.Exists(file)) throw new InvalidOperationException($"Target file {file} not copied!");
+            }
         }
+        else
+        {
+            if (_packageName != null)
+            {
+                var targetFile = Path.Combine(_packageSourceBin, $"{_packageName}.dll");;
+                var sourceFile = Path.Combine(_config.Code.MsBuildTargetDirectory!, $"{_packageName}.dll");
+                if (!File.Exists(sourceFile))
+                    throw new InvalidOperationException($"Source dll file {sourceFile} not copied!");
+                targetFile.TryCheckFileDirectory();
+                File.Copy(sourceFile, targetFile, true);
+                if (!File.Exists(targetFile)) throw new InvalidOperationException($"Package dll file {targetFile} not copied!");
+            }
+        }
+        
+        
     }
